refactor(tools):重构权限相关工具函数并优化导出方式
- 将权限提示文案从 _constant.ts 迁移至 tools/authTips.ts
- 新增 checkSelfPermission 权限检查函数
- 新增 shouldShowRequestPermissionRationale 权限请求提示判断函数- 新增 showAuthTip 授权前用户意图提示函数
- 新增 showManualAuth 用户拒绝授权后的手动设置引导函数
- 统一导出工具函数到 tools/index.ts- 更新 makePhoneCall.ts 中工具函数引用路径
- 在 package.json 中添加 tools 模块导出配置
- 移除 src/_utils.ts 文件中重复的权限处理逻辑- 优化环境变量获取逻辑，移除冗余注释
