+++ src/tools/index.ts
@@ -1,0 +1,5 @@
+export * from "./authTips";
+export * from "./checkSelfPermission";
+export * from "./shouldShowRequestPermissionRationale";
+export * from "./showAuthTip";
+export * from "./showManualAuth";
